add 1_YOLOv5_Text_Multiple_Label_Check and add except code of No File Error

+++ Data/Move/5_YOLOv5_File_Merge_For_Train.py
@@ -13,9 +13,9 @@ import argparse
 # ==============================================================
 parser = argparse.ArgumentParser(description='YOLOv5_File_Merge_For_Train')
 
-parser.add_argument('--base-path', default='/media/hi/SK Gold P31/GSC', type=str, help='옮길 데이터셋 폴더들이 존재하는 부모 폴더 경로 지정')
+parser.add_argument('--base-path', default='/media/hi/SK Gold P31/Capstone', type=str, help='옮길 데이터셋 폴더들이 존재하는 부모 폴더 경로 지정')
 parser.add_argument('--merge-path', default='/Merge_All', type=str, help='Merge 데이터셋을 저장할 폴더 경로 지정')
-parser.add_argument('--class-pathes', default=['CellPhoneV1_Move', 'CellPhoneV2_Move', 'MobilePhoneV1_Move', 'MobilePhoneV5_Move', 'PhoneDetectionV1_Move', 'PhoneDetectionV5_Move'], type=str, nargs='*', help='옮길 데이터셋 폴더명들 지정')
+parser.add_argument('--class-pathes', default=['CrowdHuman', 'Golfball_Near'], type=str, nargs='*', help='옮길 데이터셋 폴더명들 지정')
 parser.add_argument('--source-parent-pathes', default=['images', 'labels'], type=str, nargs='*', help='source 폴더 기준 부모 폴더들 경로')
 parser.add_argument('--source-child-pathes', default=['train', 'val', 'test'], type=str, nargs='*', help='source 폴더 기준 자식 폴더들 경로')
 
@@ -46,12 +46,16 @@ def get_filenames(folder_path):
 for class_path in args.class_pathes:
     for image_path in args.source_parent_pathes:
         for train_path in args.source_child_pathes:
-            # (1) source 폴더 내 파일명 추출
-            source_filenames = get_filenames(f'{args.base_path}/{class_path}/{image_path}/{train_path}')
-            # (2) target 폴더로 파일 옮기기
-            target_folder_path = f'{base_all_path}/{image_path}/{train_path}'
-            for source_filename in source_filenames:
-                source_file_path = f'{args.base_path}/{class_path}/{image_path}/{train_path}/{source_filename}'
-                print(source_file_path)
-                print(target_folder_path)
-                shutil.copy(source_file_path, target_folder_path)
+            folder_abs_path = f'{args.base_path}/{class_path}/{image_path}/{train_path}'
+            # (1) 폴더 존재하는 경우
+            if os.path.exists(folder_abs_path):
+                # 1] source 폴더 내 파일명 추출
+                source_filenames = get_filenames(folder_abs_path)
+                # 2] target 폴더로 파일 복사
+                target_folder_path = f'{base_all_path}/{image_path}/{train_path}'
+                for source_filename in source_filenames:
+                    source_file_path = f'{args.base_path}/{class_path}/{image_path}/{train_path}/{source_filename}'
+                    # 3] 복사한 파일 경로 시각화
+                    print(f'source_file_path : {source_file_path}')
+                    print(f'target_folder_path : {target_folder_path}')
+                    shutil.copy(source_file_path, target_folder_path)
